Organize tree module
